smoke particles added

+++ src/Planet.h
@@ -14,7 +14,7 @@ private:
 	Vector2 GetVersorDirection(Vector2 objectPosition); //returns versor directed from object to center of planet
 public:
 	float radius;
-	Planet(Vector2 position, double radius, double mass, Color color);
+	Planet(Vector2 position, float radius, float mass, Color color);
 
 	Vector2 GetAcceleration(Vector2 objectPosition);
 	virtual void Draw();
